requirements.txt
smartapi-python
pyotp
python-dotenv
requests
pytz
